Fixed some issues with Custom Maps & Incorrect Log Locations.
Before this fix, there were some issues with custom maps where the Raycast wouldn't find any surface as the terrain was above where the Raycast started. This should now be fixed.

As well as that, when LogAirstrikes was enabled, it would log the location of the strike before the Raycast, it now logs the exact location of the strike.

Thanks to MVP_Wendi for notifying me about the Custom Map issue.